requirements.txt
dbt==0.20.0
dbt-snowflake==0.20.0
snowflake-connector-python==2.4.6
apache-airflow-providers-snowflake==2.3.1